scraper/requirements.txt
beautifulsoup4==4.13.4
psycopg2==2.9.10
python-dotenv==1.1.0
Requests==2.32.3
